# cache_v2/cache/models/validstringresponse.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
)

// Validstringresponse validstringresponse
//
// swagger:model validstringresponse
type Validstringresponse string

// Validate validates this validstringresponse
func (m Validstringresponse) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this validstringresponse based on context it is used
func (m Validstringresponse) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}