# cache_v2/database/models/record.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// Record record
//
// swagger:model record
type Record struct {

	// employee id
	// Required: true
	EmployeeID *string `json:"employee_id"`

	// employee name
	// Required: true
	EmployeeName *string `json:"employee_name"`
}

// Validate validates this record
func (m *Record) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateEmployeeID(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateEmployeeName(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *Record) validateEmployeeID(formats strfmt.Registry) error {

	if err := validate.Required("employee_id", "body", m.EmployeeID); err != nil {
		return err
	}

	return nil
}

func (m *Record) validateEmployeeName(formats strfmt.Registry) error {

	if err := validate.Required("employee_name", "body", m.EmployeeName); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this record based on context it is used
func (m *Record) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *Record) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *Record) UnmarshalBinary(b []byte) error {
	var res Record
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}